Extract AtomicStack
