+++ tokio-timer/src/timer/entry.rs
@@ -8,7 +8,7 @@ use futures::task::AtomicTask;
 use std::cell::UnsafeCell;
 use std::ptr;
 use std::sync::{Arc, Weak};
-use std::sync::atomic::{AtomicBool, AtomicPtr};
+use std::sync::atomic::AtomicBool;
 use std::sync::atomic::Ordering::SeqCst;
 use std::time::Instant;
 use std::u64;
@@ -56,12 +56,12 @@ pub(crate) struct Entry {
 
     /// True when the entry is queued in the "process" stack. This value
     /// is set before pushing the value and unset after popping the value.
-    queued: AtomicBool,
+    pub(super) queued: AtomicBool,
 
     /// Next entry in the "process" linked list.
     ///
     /// Represents a strong Arc ref.
-    next_atomic: UnsafeCell<*mut Entry>,
+    pub(super) next_atomic: UnsafeCell<*mut Entry>,
 
     /// When the entry expires, relative to the `start` of the timer
     /// (Inner::start). This is only used by the timer.
@@ -91,27 +91,11 @@ pub(crate) struct Entry {
     pub(super) prev_stack: UnsafeCell<*const Entry>,
 }
 
-/// A stack of `Entry` nodes
-#[derive(Debug)]
-pub(crate) struct AtomicStack {
-    /// Stack head
-    head: AtomicPtr<Entry>,
-}
-
-/// Entries that were removed from the stack
-#[derive(Debug)]
-pub(crate) struct AtomicStackEntries {
-    ptr: *mut Entry,
-}
-
 /// Flag indicating a timer entry has elapsed
 const ELAPSED: u64 = 1 << 63;
 
 /// Flag indicating a timer entry has reached an error state
 const ERROR: u64 = u64::MAX;
-
-/// Used to indicate that the timer has shutdown.
-const SHUTDOWN: *mut Entry = 1 as *mut _;
 
 // ===== impl Entry =====
 
@@ -344,104 +328,3 @@ impl Drop for Entry {
 
 unsafe impl Send for Entry {}
 unsafe impl Sync for Entry {}
-
-// ===== impl AtomicStack =====
-
-impl AtomicStack {
-    pub fn new() -> AtomicStack {
-        AtomicStack { head: AtomicPtr::new(ptr::null_mut()) }
-    }
-
-    /// Push an entry onto the stack.
-    ///
-    /// Returns `true` if the entry was pushed, `false` if the entry is already
-    /// on the stack, `Err` if the timer is shutdown.
-    pub fn push(&self, entry: &Arc<Entry>) -> Result<bool, Error> {
-        // First, set the queued bit on the entry
-        let queued = entry.queued.fetch_or(true, SeqCst).into();
-
-        if queued {
-            // Already queued, nothing more to do
-            return Ok(false);
-        }
-
-        let ptr = Arc::into_raw(entry.clone()) as *mut _;
-
-        let mut curr = self.head.load(SeqCst);
-
-        loop {
-            if curr == SHUTDOWN {
-                // Don't leak the entry node
-                let _ = unsafe { Arc::from_raw(ptr) };
-
-                return Err(Error::shutdown());
-            }
-
-            // Update the `next` pointer. This is safe because setting the queued
-            // bit is a "lock" on this field.
-            unsafe {
-                *(entry.next_atomic.get()) = curr;
-            }
-
-            let actual = self.head.compare_and_swap(curr, ptr, SeqCst);
-
-            if actual == curr {
-                break;
-            }
-
-            curr = actual;
-        }
-
-        Ok(true)
-    }
-
-    /// Take all entries from the stack
-    pub fn take(&self) -> AtomicStackEntries {
-        let ptr = self.head.swap(ptr::null_mut(), SeqCst);
-        AtomicStackEntries { ptr }
-    }
-
-    /// Drain all remaining nodes in the stack and prevent any new nodes from
-    /// being pushed onto the stack.
-    pub fn shutdown(&self) {
-        // Shutdown the processing queue
-        let ptr = self.head.swap(SHUTDOWN, SeqCst);
-
-        // Let the drop fn of `AtomicStackEntries` handle draining the stack
-        drop(AtomicStackEntries { ptr });
-    }
-}
-
-// ===== impl AtomicStackEntries =====
-
-impl Iterator for AtomicStackEntries {
-    type Item = Arc<Entry>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.ptr.is_null() {
-            return None;
-        }
-
-        // Convert the pointer to an `Arc<Entry>`
-        let entry = unsafe { Arc::from_raw(self.ptr) };
-
-        // Update `self.ptr` to point to the next element of the stack
-        self.ptr = unsafe { (*entry.next_atomic.get()) };
-
-        // Unset the queued flag
-        let res = entry.queued.fetch_and(false, SeqCst);
-        debug_assert!(res);
-
-        // Return the entry
-        Some(entry)
-    }
-}
-
-impl Drop for AtomicStackEntries {
-    fn drop(&mut self) {
-        while let Some(entry) = self.next() {
-            // Flag the entry as errored
-            entry.error();
-        }
-    }
-}
